requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pinecone
python-dotenv
langchain-text-splitters
google-genai
